Database refactor, fixes defect in bills to send to constituents
When sending the tally mail, bills that the constituent had just voted
on were not being returned for the email. (The logic only appreciated
the last time congress addressed the bill.)

Upserts were being silently lost in the background due to open connections
not committing their transactions. It's odd behavior given the
documentation on psycopg2.

* Refactored the DB class to use `with` contexts
* Created a connection class to handle things like a test environment
* Refactored code that depends on the database to 1) inherit its db
  dependencies so fetching and writing to the db is done through `self`
  and 2) does not have to manage commits/rollbacks on its own
* updates tests