Saves categories of bills to database and relates them
We're in the early stages of wanting to show bills to users sorted by
how likely they are to vote for the bill. A naive, first approach,
to solving this problem will include determining if a bill contains
a category they've voted on before.

Fixes PEP8 issues in a few files as well.

Signed-off-by: Bryce McDonnell <bryce@bridgetownint.com>